Move local_addr_ready test into the tcp test file
